requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
exoscale==0.10.0
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
requests-exoscale-auth==1.1.2
rich==13.7.1
urllib3==2.2.1
